# cmd/api/docs/swagger.yaml
definitions:
...
  output.OrderItemOutputDTO:
    properties:
      amount_value:
        type: string
      fk_resale_order_id:
        type: string
      id:
        type: string
      name:
        type: string
      quantity:
        type: integer
      shipping_code:
        type: string
      shipping_status:
        type: string
      sku:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/app/users/{cpf}/orders/{order_id}/items:
    get:
      consumes:
      - application/json
      description: Returns all items of a specific order identified by user CPF and
        order ID
      parameters:
      - description: User CPF (digits only, 11 characters)
        in: path
        name: cpf
        required: true
        type: string
      - description: Order ID (UUID)
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order items retrieved successfully
          schema:
            items:
              $ref: '#/definitions/output.OrderItemOutputDTO'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Order not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List order items by CPF and order ID
      tags:
      - Resale
  /v1/app/users/{cpf}/orders/{order_id}/items/{item_id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancels a specific item of an order identified by user CPF, order
        ID and item ID
      parameters:
      - description: User CPF (digits only, 11 characters)
        in: path
        name: cpf
        required: true
        type: string
      - description: Order ID (UUID)
        in: path
        name: order_id
        required: true
        type: string
      - description: Order Item ID (UUID)
        in: path
        name: item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Item cancelled successfully
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Order or item not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel an order item
      tags:
      - Resale
...
swagger: "2.0"